feat: enforce ticket size int type

+++ backend/src/schema/setting.ts
@@ -2,7 +2,7 @@ import * as z from "zod";
 
 export const SettingSchema = z.object({
   duplicate_threshold: z.number(),
-  max_ticket_size: z.number(),
+  max_ticket_size: z.number().int(),
 });
 
 export type Setting = z.infer<typeof SettingSchema>;
